iox-#1614 Rename concurrent::Buffer to containers::UninitializedArray
