+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/containers/uninitialized_array.inl
@@ -1,4 +1,5 @@
 // Copyright (c) 2020 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,54 +15,54 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef IOX_HOOFS_CONCURRENT_LOCKFREE_QUEUE_BUFFER_INL
-#define IOX_HOOFS_CONCURRENT_LOCKFREE_QUEUE_BUFFER_INL
+#ifndef IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
+#define IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
 
 #include "iceoryx_hoofs/containers/uninitialized_array.hpp"
 
 namespace iox
 {
-namespace concurrent
+namespace containers
 {
 template <typename ElementType, uint64_t Capacity, typename index_t>
-ElementType& Buffer<ElementType, Capacity, index_t>::operator[](const index_t index) noexcept
-{
-    return *toPtr(index);
-}
-
-template <typename ElementType, uint64_t Capacity, typename index_t>
-const ElementType& Buffer<ElementType, Capacity, index_t>::operator[](const index_t index) const noexcept
+ElementType& UnitializedArray<ElementType, Capacity, index_t>::operator[](const index_t index) noexcept
 {
     return *toPtr(index);
 }
 
 template <typename ElementType, uint64_t Capacity, typename index_t>
-ElementType* Buffer<ElementType, Capacity, index_t>::ptr(const index_t index) noexcept
+const ElementType& UnitializedArray<ElementType, Capacity, index_t>::operator[](const index_t index) const noexcept
+{
+    return *toPtr(index);
+}
+
+template <typename ElementType, uint64_t Capacity, typename index_t>
+ElementType* UnitializedArray<ElementType, Capacity, index_t>::ptr(const index_t index) noexcept
 {
     return toPtr(index);
 }
 
 template <typename ElementType, uint64_t Capacity, typename index_t>
-const ElementType* Buffer<ElementType, Capacity, index_t>::ptr(const index_t index) const noexcept
+const ElementType* UnitializedArray<ElementType, Capacity, index_t>::ptr(const index_t index) const noexcept
 {
     return toPtr(index);
 }
 
 template <typename ElementType, uint64_t Capacity, typename index_t>
-uint64_t Buffer<ElementType, Capacity, index_t>::capacity() const noexcept
+uint64_t UnitializedArray<ElementType, Capacity, index_t>::capacity() const noexcept
 {
     return Capacity;
 }
 
 template <typename ElementType, uint64_t Capacity, typename index_t>
-ElementType* Buffer<ElementType, Capacity, index_t>::toPtr(index_t index) const noexcept
+ElementType* UnitializedArray<ElementType, Capacity, index_t>::toPtr(index_t index) const noexcept
 {
     auto ptr = &(m_buffer[index * sizeof(ElementType)]);
     // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast, cppcoreguidelines-pro-type-const-cast) type erasure
     return reinterpret_cast<ElementType*>(const_cast<byte_t*>(ptr));
 }
 
-} // namespace concurrent
+} // namespace containers
 } // namespace iox
 
-#endif // IOX_HOOFS_CONCURRENT_LOCKFREE_QUEUE_BUFFER_INL
+#endif // IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
